finito

+++ src/main/java/marketplace/ProjetJ2EE_SpringBoot/controllers/EditController.java
@@ -1,9 +1,11 @@
 package marketplace.ProjetJ2EE_SpringBoot.controllers;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import marketplace.ProjetJ2EE_SpringBoot.model.Client;
 import marketplace.ProjetJ2EE_SpringBoot.model.Produit;
 import marketplace.ProjetJ2EE_SpringBoot.service.DetailCommandeService;
 import marketplace.ProjetJ2EE_SpringBoot.service.ProduitService;
-import org.springframework.boot.Banner;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -18,7 +20,10 @@ public class EditController {
         this.produitService = produitService; this.detailCommandeService = detailCommandeService;}
 
     @GetMapping("/edit/{id}")
-    public String editProductPage(@PathVariable String id, Model model) {
+    public String editProductPage(@PathVariable String id, Model model, HttpServletRequest request) {
+        if(!isConnected(request.getSession())){
+            return "redirect:/account";
+        }
         Produit product = produitService.findProduitById(Integer.parseInt(id));
 
         model.addAttribute("product", product);
@@ -33,24 +38,41 @@ public class EditController {
             @RequestParam String description,
             @RequestParam String stock,
             @RequestParam String imageUrl,
-            Model model
+            HttpServletRequest request
     ) {
-        produitService.saveProduit(Integer.parseInt(id),nom,Double.parseDouble(price),description,Integer.parseInt(stock),imageUrl);
+        if(!isConnected(request.getSession())){
+            return "redirect:/account";
+        }
+        Produit produit = new Produit();
+        produit.setNom(nom);
+        produit.setId(Integer.parseInt(id));
+        produit.setPrix(Double.parseDouble(price));
+        produit.setDescription(description);
+        produit.setStock(Integer.parseInt(stock));
+        produit.setImage(imageUrl);
+
+        produitService.saveProduit(produit);
         return "redirect:/product/edit/" + id ;
     }
 
     @GetMapping("/delete/{idProduit}")
-    public String supprimerProduit(@PathVariable int idProduit, Model model) {
+    public String supprimerProduit(@PathVariable int idProduit, Model model, HttpServletRequest request) {
+        if(!isConnected(request.getSession())){
+            return "redirect:/account";
+        }
+
         detailCommandeService.mettreAJourProduitId(idProduit);
 
         int rowsAffected = produitService.deleteProduit(idProduit);
         model.addAttribute("rowsAffected", rowsAffected);
-        return "/deleteProduct";
+        return "deleteProduct";
     }
 
     @GetMapping("/create")
-    public String creerProduit(){
-
+    public String creerProduit(HttpSession session){
+        if(!isConnected(session)){
+            return "redirect:/account";
+        }
         return "addProduct";
     }
 
@@ -61,8 +83,11 @@ public class EditController {
             @RequestParam String description,
             @RequestParam int stock,
             @RequestParam String image,
-            Model model
+            HttpServletRequest request
     ) {
+        if(!isConnected(request.getSession())){
+            return "redirect:/account";
+        }
         Produit nouveauProduit = new Produit();
         nouveauProduit.setDescription(description);
         nouveauProduit.setPrix(prix);
@@ -72,5 +97,10 @@ public class EditController {
         produitService.saveProduit(nouveauProduit);
 
         return "redirect:/editProducts";
+    }
+
+    private boolean isConnected(HttpSession session) {
+        Client client = (Client) session.getAttribute("client");
+        return ((client != null && !client.getDroit().equals("aucun")) || (session.getAttribute("role") != null && session.getAttribute("role").equals("admin")));
     }
 }
